Big model reorganization
What was previously and pretentiosly called the "model" is now called "schema"
and continues to include rather low-level database functions that build directly
on DBIx::Simple. The new model classes will work on Ashafix::Result::* objects
and do pretty much all data validation, i.e. interfacing the business logic to
the schema.

This is still rather broken as I've only started to separate the different
levels of logic out of the controllers.